style: добавлены конфиги prettier,eslint,ts,editor
