+++ .eslintrc.js
@@ -1,0 +1,57 @@
+module.exports = {
+    extends: ['airbnb-base', 'plugin:@typescript-eslint/eslint-recommended', 'plugin:@typescript-eslint/recommended'],
+    parser: '@typescript-eslint/parser',
+    plugins: ['import', 'prettier'],
+    rules: {
+        '@typescript-eslint/no-empty-function': 'off',
+        '@typescript-eslint/no-empty-interface': 'off',
+        'no-restricted-syntax': 'off', // В for...of циклах ничего плохого нет
+        'spaced-comment': ['error', 'always', { markers: ['/'] }], /// разрешаем ts-require directive
+        'comma-dangle': ['error', 'always-multiline'],
+        'arrow-parens': ['error', 'always'],
+        indent: 'off',
+        'max-len': [
+            'error',
+            120,
+            2,
+            {
+                ignoreUrls: true,
+                ignoreComments: false,
+                ignoreRegExpLiterals: true,
+                ignoreStrings: true,
+                ignoreTemplateLiterals: true,
+            },
+        ],
+        'implicit-arrow-linebreak': 'off',
+        'no-plusplus': 'off',
+        'max-classes-per-file': 'off',
+        'operator-linebreak': 'off',
+        'object-curly-newline': 'off',
+        'class-methods-use-this': 'off',
+        'no-confusing-arrow': 'off',
+        'function-paren-newline': 'off',
+        'no-param-reassign': 'off',
+        'no-shadow': 'warn',
+        'space-before-function-paren': 'off',
+        'consistent-return': 'off',
+        'prettier/prettier': 'error',
+
+        '@typescript-eslint/explicit-function-return-type': 'off',
+
+        'import/prefer-default-export': 'off', // https://humanwhocodes.com/blog/2019/01/stop-using-default-exports-javascript-module/
+        'import/order': [
+            'error',
+            {
+                groups: [['builtin', 'external'], 'internal', 'parent', 'sibling', 'index'],
+                'newlines-between': 'always',
+            },
+        ],
+        'import/no-unresolved': 'off',
+        'import/extensions': 'off',
+        'import/no-extraneous-dependencies': ['off'], // можно включить тока нужно резолвы разрулить
+        'arrow-body-style': 'off',
+        'padding-line-between-statements': 'off',
+        'no-unused-expressions': 'off',
+        camelcase: 'off',
+    },
+};
