+++ .prettierrc
@@ -1,0 +1,10 @@
+{
+  "parser": "typescript",
+  "arrowParens": "always",
+  "printWidth": 120,
+  "endOfLine": "auto",
+  "semi": true,
+  "singleQuote": true,
+  "tabWidth": 4,
+  "trailingComma": "all"
+}
