+++ package.json
@@ -4,8 +4,10 @@
   "description": "Tiny helpers to make scenario for Salute Assistants family on Nodejs",
   "author": "SberDevices Frontend Team <sberdevices.frontend@gmail.com>",
   "license": "Sber Public License at-nc-sa v.2",
-  "main": "index.js",
+  "main": "dist/index.d.ts",
   "scripts": {
+    "build": "rm -rf dist && tsc",
+    "lint": "eslint --ext .js,.ts ./src",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -15,5 +17,17 @@
   "bugs": {
     "url": "https://github.com/sberdevices/salutejs/issues"
   },
-  "homepage": "https://github.com/sberdevices/salutejs#readme"
+  "homepage": "https://github.com/sberdevices/salutejs#readme",
+  "devDependencies": {
+    "@types/node": "14.14.25",
+    "@typescript-eslint/eslint-plugin": "4.14.2",
+    "@typescript-eslint/parser": "4.14.2",
+    "eslint": "7.19.0",
+    "eslint-config-airbnb-base": "14.2.1",
+    "eslint-config-prettier": "7.2.0",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-prettier": "3.3.1",
+    "prettier": "2.2.1",
+    "typescript": "4.1.3"
+  }
 }
