+++ src/index.ts
@@ -1,0 +1,2 @@
+export { NLPRequest } from './types/request';
+export { NLPResponse } from './types/response';
